EXAMPLES: renamed header file
