+++ examples/Makefile.am
@@ -6,10 +6,10 @@
 #
 
 examplesdir = $(pkgdatadir)/examples
-dist_examples_DATA =  \
-	ucx_hello_world.h \
-	ucp_hello_world.c \
-	uct_hello_world.c \
+dist_examples_DATA =   \
+	hello_world_util.h \
+	ucp_hello_world.c  \
+	uct_hello_world.c  \
 	ucp_client_server.c
 
 if HAVE_CUDA
