+++ examples/hello_world_util.h
